+++ public/ckeditor/styles.js
@@ -44,7 +44,11 @@ CKEDITOR.stylesSet.add( 'default', [
 			border: '1px solid #ccc'
 		}
 	},
-
+	{
+		name: 'Image Container',
+		element: 'div',
+		attributes: {'class':'content-image-container'}
+	},	
 	/* Inline styles */
 
 	// These are core styles available as toolbar buttons. You may opt enabling
@@ -91,6 +95,12 @@ CKEDITOR.stylesSet.add( 'default', [
 		name: 'Styled Image (right)',
 		element: 'img',
 		attributes: { 'class': 'right' }
+	},
+
+	{
+		name: 'Footer / References',
+		element: 'p',
+		attributes: { 'class': 'references' }
 	},
 
 	{