+++ resources/views/layouts/master.blade.php
@@ -35,6 +35,8 @@
 		<link rel="stylesheet" type="text/css" href="{{ mix('/css/style.css') }}" />
 		{!! Html::style('css/slick-theme.css') !!}
 		{!! Html::style('css/magnific-popup.css') !!}
+		{!! Html::style('css/plugins/slick/slick.css') !!}
+		{!! Html::style('css/content-slick.css') !!}
 
 		{{-- JS --}}
 		<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.10.2/jquery.min.js"></script>
@@ -105,14 +107,28 @@ height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
 	</div>
 
 	<!-- Core JavaScript Files -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.3/jquery.easing.min.js"></script>
     <script src="https://unpkg.com/isotope-layout@3/dist/isotope.pkgd.min.js"></script>
 	<script src="https://unpkg.com/flickity@1.1/dist/flickity.pkgd.min.js"></script>
 	<script src="https://unpkg.com/infinite-scroll@3.0.6/dist/infinite-scroll.pkgd.min.js"></script>
-    {!! Html::script('js/bootstrap.min.js') !!}
+	{!! Html::script('js/bootstrap.min.js') !!}
+	{!! Html::script('js/slick.js') !!}
+	{!! Html::script('js/jquery.easing.min.js') !!}
     @include('components.searchbar_js')
     <script src="{!! asset_timed('js/webumenia.js') !!}"></script>
+	<script type="text/javascript">
+		$(document).ready(function(){
+			$('.content-slick-images').each( function() {
 
+				const $slick = $(this);
+				$slick.slick({
+					slide: 'p, a, img',
+					lazyLoad: 'progressive',
+					variableWidth: true,
+					infinite: false,
+				});
+			});
+		});
+	</script>
 
 	<!-- Content -->
 	@yield('javascript')